move code into a folder hierarchy that matches the code hierarchy
